Removing Warnings  Unecesasry Casts  Local Variable Never Read  Import Never Used

+++ src_tests/ServidorAplicacao/Servicos/teacher/CountMetadatasByExecutionCourseTest.java
@@ -88,7 +88,7 @@ public class CountMetadatasByExecutionCourseTest extends ServiceNeedsAuthenticat
 			Criteria criteria = new Criteria();
 			criteria.addEqualTo("keyExecutionCourse", args[0]);
 			Query queryCriteria = new QueryByCriteria(Metadata.class, criteria);
-			int metadataNumber = (int) broker.getCount(queryCriteria);
+			int metadataNumber = broker.getCount(queryCriteria);
 			broker.close();
 			if (count.intValue() != metadataNumber)
 				fail("CountMetadatasByExecutionCourseTest " + "wrong number of metadatas");
